requirements-voice.txt
# LiveKit
livekit>=0.17.0
livekit-agents>=0.12.0
livekit-plugins-silero>=0.7.0

# STT
faster-whisper>=1.1.0
# Note: For GPU support, install ctranslate2 with CUDA:
# pip install ctranslate2 --extra-index-url https://download.pytorch.org/whl/cu121

# TTS
kokoro>=0.9.2  # Kokoro-82M: lightweight, high-quality neural TTS (default)
piper-tts==1.2.0  # Piper: local CPU-based TTS (fallback)
edge-tts>=6.1.0  # Microsoft Edge neural TTS (cloud fallback)
pydub>=0.25.0  # Audio conversion for Edge TTS
soundfile>=0.12.0  # Audio I/O for Kokoro

# HTTP
httpx>=0.26.0
aiohttp>=3.10.0
websockets>=12.0

# Redis
redis==5.0.1

# Utils
python-dotenv==1.0.1
structlog==24.1.0
numpy==1.26.3
pydantic-settings>=2.1.0
